% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_extrapolation_info.R
\name{make_extrapolation_info}
\alias{make_extrapolation_info}
\title{Build extrapolation grid}
\usage{
make_extrapolation_info(strata.limits, stream_info = NULL, zone = NA,
  flip_around_dateline = TRUE, ...)
}
\arguments{
\item{strata.limits}{an input for determining stratification of indices (see example script)}

\item{stream_info}{data frame with 3 columns: latitude, longitude, and child node of stream segment (upstream node) associated with each observation}
}
\value{
Tagged list used in other functions
\describe{
  \item{a_el}{The area associated with each extrapolation grid cell (rows) and strata (columns)}
  \item{Data_Extrap}{A data frame describing the extrapolation grid}
  \item{zone}{the zone used to convert Lat-Long to UTM by PBSmapping package}
  \item{flip_around_dateline}{a boolean stating whether the Lat-Long is flipped around the dateline during conversion to UTM}
  \item{Area_km2_x}{the area associated with each row of Data_Extrap, in units square-kilometers}
}
}
\description{
\code{make_extrapolation_data} builds an object used to determine areas to extrapolation densities to when calculating indices
}
