\name{nz_longfin_eel.rda}
\alias{nz_longfin_eel}
\docType{data}
\title{
New Zealand longfin eel data from the Waitaki River catchment
}
\description{
A list with 2 elements, `observations` and `network`. Observations include the presence or absence of longfin eel at each sampling location and time period, with a small, normally distributed random value with mean = 0 and standard deviation 0.00001 added to all encounters. The network includes all child (upstream) and parent(downstream) nodes within the Waitaki River catchment. Upper portions of the catchment were excluded to include the lower region with observations (excluding the upper region with no observations, and thus no information on density). 
}
\usage{data("nz_longfin_eel", package="FishStatsUtils-streams")}
\source{
New Zealand River Environment Classification, contact Shannan Crow (Shannan.Crow@niwa.co.nz) or Simon Hoyle (Simon.Hoyle@niwa.co.nz)
}
\keyword{datasets}
