\name{or_siletz_coho.rda}
\alias{or_siletz_coho}
\docType{data}
\title{
Oregon Siletz watershed coho data
}
\description{
A list with 2 elements, `observations` and `network`. Observations include the density of juvenile and spawner coho salmon at each sampling location and time period. The network includes all child (upstream) and parent(downstream) nodes within the Siletz watershed. 
}
\usage{data("or_siletz_coho", package="StreamUtils")}
\source{
Oregon Department of Fish and Wildlife
}
\keyword{datasets}
