% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_maps.r
\name{plot_maps}
\alias{plot_maps}
\title{Plot standard maps}
\usage{
plot_maps(plot_set = 3, Report, Sdreport = NULL, TmbData = NULL,
  savedir = paste0(getwd(), "/"), category_names = NULL,
  cex_network = 1, X_xtp, ...)
}
\arguments{
\item{plot_set}{integer-vector defining plots to create
\describe{
  \item{plot_set=1}{Probability of encounter/no-encounter}
  \item{plot_set=2}{Log-expected positive catch rate}
  \item{plot_set=3}{Log-predicted density (product of encounter probability and positive catch rates)}
  \item{plot_set=4}{Log-positive catch rates (rescaled)}
  \item{plot_set=5}{Log-predicted density (rescaled)}
  \item{plot_set=6}{Spatio-temporal variation in encounter probability}
  \item{plot_set=7}{Spatio-temporal variation in log-positive catch rates}
  \item{plot_set=8}{Linear predictor for encounter probability}
  \item{plot_set=9}{Linear predictor for positive catch rates}
  \item{plot_set=10}{Coefficient of variation for predicted density (available only if \code{Data_Fn(...,Options=c('SD_site_logdensity'=1,...))}}
  \item{plot_set=11}{Covariates that are included in the model}
  \item{plot_set=12}{Total biomass across all categories (only useful in a multivariate model)}
  \item{plot_set=13}{Covariate effects on encounter probability}
  \item{plot_set=14}{Covariate effects on positive catch rates}
  \item{plot_set=15}{Covariate effect on encounter probability by covariate}
}}

\item{Report}{tagged list of outputs from TMB model via \code{Obj$report()}}

\item{Sdreport}{Standard deviation outputs from TMB model via \code{sdreport(Obj)}}

\item{savedir}{Directory (absolute path) and base for filenames of plots}

\item{category_names}{character vector specifying names for different categories (only used for R package \code{VAST})}

\item{cex_network}{point size for network}

\item{...}{arguments passed to \code{PlotMap_Fn}}

\item{Year_Set}{total years}

\item{Years2Include}{years with data}
}
\value{
Mat_xt a matrix (rows: modeled knots; column: modeled year) for plotted output of last element of \code{plot_set}
}
\description{
\code{plot_maps} plots a standard set of diagnostic maps for stream network models
}
