% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Plot stream network and optional observations}
\usage{
plot_network(Extrapolation_List, Spatial_List, Data, Data_Geostat,
  observations = TRUE, arrows = TRUE, root = TRUE, PlotDir = NULL,
  ...)
}
\arguments{
\item{Extrapolation_List}{Output from \code{Prepare_Extrapolation_Data_Fn}}

\item{Spatial_List}{Output from \code{FishStatsUtils::make_spatial_list}}

\item{Data}{output of VAST::Data_Fn which includes parent_s and child_s}

\item{Data_Geostat}{data-frame of data (with columns 'Year', 'Lon', 'Lat' or 'E_km', 'N_km' at a minimum)}

\item{observations}{default TRUE to include observations on the figure, FALSE to show only network}

\item{arrows}{default TRUE to show directional connections between nodes, but FALSE may prevent crash if there are too many points.}

\item{root}{default TRUE to show root nodes, FALSE in case there are other root nodes that are not meaningful.}

\item{PlotDir}{Directory for plots}

\item{...}{addition inputs to \code{plot}}
}
\value{
Figure plotting stream network and observations
}
\description{
\code{plot_network} generates figure showing network nodes, directional arrows, and observations
}
