% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Plot Pearson residuals on map}
\usage{
plot_residuals(Report, Q, savedir = getwd(), plot_type = c(1, 2), ...)
}
\arguments{
\item{Q}{Output from \code{QQ_Fn}}

\item{savedir}{directory to use when saving results}

\item{plot_type}{1 == encounter probability, 2 == catch rates}

\item{...}{arguments passed to \code{PlotMap_Fn}}
}
\value{
A tagged list of Pearson residuals
\describe{
  \item{Q1_xt}{Matrix of average residuals for encounter/non-encounter component by site \code{x} and year \code{t}}
  \item{Q2_xt}{Matrix of average residuals for positive-catch-rate component by site \code{x} and year \code{t}}
}
}
\description{
\code{plot_residuals} shows average Pearson residual for every knot for encounter probability and positive catch rate components
}
