\name{StreamUtils}
\alias{StreamUtils-package}
\alias{StreamUtils}
\docType{package}
\title{StreamUtils}
\description{
StreamUtils supplements the utilities included in FishStatsUtils for spatio-temporal models (VAST) for stream networks
}
\details{
Please see examples folder on github \url{http://github.com/merrillrudd/StreamUtils/tree/master/examples} for up-to-date examples
}
